price changed to getDouble, item codes updated

+++ src/model/Cookie.java
@@ -42,7 +42,7 @@ public class Cookie implements SnackComponent {
                      sugars = result.getInt("sugars");
                     carbohydrates = result.getInt("carbohydrates");
                     calories= result.getInt("calories");
-                    price = result.getInt("price");
+                    price = result.getDouble("price");
                     count = result.getInt("count");
                 }
