+++ src/model/GreekYogurt.java
@@ -41,7 +41,7 @@ public class GreekYogurt implements SnackComponent{
                      sugars = result.getInt("sugars");
                     carbohydrates = result.getInt("carbohydrates");
                     calories= result.getInt("calories");
-                    price = result.getInt("price");
+                    price = result.getDouble("price");
                     count = result.getInt("count");
                 }
